refactor shared to internal
