+++ internal/identity/aws.go
@@ -8,35 +8,33 @@ import (
 	"github.com/tink-crypto/tink-go/v2/keyset"
 	"github.com/tink-crypto/tink-go/v2/signature"
 	"github.com/tink-crypto/tink-go/v2/tink"
-
-	"github.com/TBD54566975/ftl/backend/controller/identity/api"
 )
 
-var _ api.IdentitySigner = &KMSIdentitySigner{}
+var _ Signer = &KMSIdentitySigner{}
 
 type KMSIdentitySigner struct {
 	signer tink.Signer
 }
 
-func (k KMSIdentitySigner) Sign(data []byte) (*api.SignedData, error) {
+func (k KMSIdentitySigner) Sign(data []byte) (*SignedData, error) {
 	bytes, err := k.signer.Sign(data)
 	if err != nil {
 		return nil, fmt.Errorf("failed to sign data: %w", err)
 	}
 
-	return &api.SignedData{
+	return &SignedData{
 		Data:      data,
 		Signature: bytes,
 	}, nil
 }
 
-var _ api.IdentityVerifier = &KMSIdentityVerifier{}
+var _ Verifier = &KMSIdentityVerifier{}
 
 type KMSIdentityVerifier struct {
 	verifier tink.Verifier
 }
 
-func (k KMSIdentityVerifier) Verify(signedData api.SignedData) error {
+func (k KMSIdentityVerifier) Verify(signedData SignedData) error {
 	err := k.verifier.Verify(signedData.Signature, signedData.Data)
 	if err != nil {
 		return fmt.Errorf("failed to verify signature: %w", err)
@@ -45,13 +43,13 @@ func (k KMSIdentityVerifier) Verify(signedData api.SignedData) error {
 	return nil
 }
 
-var _ api.KeyPair = &TinkKeyPair{}
+var _ KeyPair = &TinkKeyPair{}
 
 type TinkKeyPair struct {
 	keysetHandle keyset.Handle
 }
 
-func (t TinkKeyPair) Signer() (api.IdentitySigner, error) {
+func (t TinkKeyPair) Signer() (Signer, error) {
 	signer, err := signature.NewSigner(&t.keysetHandle)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create signer: %w", err)
@@ -62,7 +60,7 @@ func (t TinkKeyPair) Signer() (api.IdentitySigner, error) {
 	}, nil
 }
 
-func (t TinkKeyPair) Verifier() (api.IdentityVerifier, error) {
+func (t TinkKeyPair) Verifier() (Verifier, error) {
 	public, err := t.keysetHandle.Public()
 	if err != nil {
 		return nil, fmt.Errorf("failed to get public keyset from keyset handle: %w", err)
